Api for employee dashboard leaderboard

+++ src/main/java/com/perkpal/dto/EmployeeLeaderBoardDto.java
@@ -11,5 +11,6 @@ public class EmployeeLeaderBoardDto {
     private String fullName;
     private String departmentName;
     private String photoUrl;
-    private Long totalPoints;
+    private Double totalPoints;
 }
+
